Refactoring project structure (air-condition module)
