+++ src/app/air-condition/actions/stations.actions.ts
