+++ src/app/air-condition/components/map/map.component.css
@@ -1,5 +1,5 @@
 agm-map {
-  height: 80vh;
+  height: 100%;
 }
 
 .marker-title {
