+++ src/app/air-condition/components/stations/stations.component.html
