--- src/app/air-condition/map/components/map/map.component.html
@@ -1,23 +1,0 @@
-<div class="fetching" *ngIf="(isFetching$ | async)">
-  <h2 class="fetch-title">Feching map data</h2>
-  <mat-spinner></mat-spinner>
-</div>
-<agm-map [latitude]="lat" [longitude]="lng" [zoom]="zoom" [styles]="styleArray">
-  <ng-container *ngFor="let station of (stationsList | async)">
-    <agm-marker
-      [latitude]="station.gegrLat"
-      [longitude]="station.gegrLon"
-      [iconUrl]="markerIcon"
-      (markerClick)="this.handleMarkerClick(station.stationName, station.id)"
-    >
-      <agm-snazzy-info-window
-        [backgroundColor]="000000"
-        [maxWidth]="600"
-        [maxHeight]="600"
-        [closeWhenOthersOpen]="true"
-      >
-        <ng-template> <anms-stations-modal></anms-stations-modal> </ng-template>
-      </agm-snazzy-info-window>
-    </agm-marker>
-  </ng-container>
-</agm-map>
