+++ src/app/air-condition/models/stationsModels.ts
