+++ src/app/air-condition/reducers/stations.reducer.ts
