+++ src/app/app.component.scss
@@ -49,6 +49,7 @@ mat-sidenav-container {
       flex: 1 0 auto;
       margin-top: 64px;
       overflow: hidden;
+      height: 100%;
     }
 
     .footer {
